requirements.txt
Flask
selenium
webdriver-manager
click
pikepdf
Pillow